+++ app/Http/Controllers/ExtensionPrograms/ExpertServices/AcademicController.php
@@ -23,6 +23,8 @@ class AcademicController extends Controller
      */
     public function index()
     {
+        $this->authorize('viewAny', ExpertServiceAcademic::class);
+
         $expertServicesAcademic = ExpertServiceAcademic::where('user_id', auth()->id())
                                         ->join('dropdown_options', 'dropdown_options.id', 'expert_service_academics.classification')
                                         ->select('expert_service_academics.*', 'dropdown_options.name as classification')
@@ -38,6 +40,8 @@ class AcademicController extends Controller
      */
     public function create()
     {
+        $this->authorize('create', ExpertServiceAcademic::class);
+
         if(ExtensionProgramForm::where('id', 3)->pluck('is_active')->first() == 0)
             return view('inactive');
         $expertServiceAcademicFields = DB::select("CALL get_extension_program_fields_by_form_id(3)");
@@ -55,6 +59,7 @@ class AcademicController extends Controller
      */
     public function store(Request $request)
     {
+        $this->authorize('create', ExpertServiceAcademic::class);
         if(ExtensionProgramForm::where('id', 3)->pluck('is_active')->first() == 0)
             return view('inactive');
 
@@ -64,12 +69,12 @@ class AcademicController extends Controller
             'from' => 'required|date',
             'to' => 'required|date|after_or_equal:from',
             // 'publication_or_audio_visual' => '',
-            // 'copyright_no' => '',
+            'copyright_no' => 'required',
             // 'indexing' => '',
             'level' => 'required',
             'college_id' => 'required',
             'department_id' => 'required',
-            'description' => 'required',
+            // 'description' => 'required',
         ]);
 
         $input = $request->except(['_token', '_method', 'document', 'college_id']);
@@ -111,23 +116,19 @@ class AcademicController extends Controller
      */
     public function show(ExpertServiceAcademic $expert_service_in_academic)
     {
+        $this->authorize('view', ExpertServiceAcademic::class);
+
         if(ExtensionProgramForm::where('id', 3)->pluck('is_active')->first() == 0)
             return view('inactive');
-        $expertServiceAcademicDocuments = ExpertServiceAcademicDocument::where('expert_service_academic_id', $expert_service_in_academic->id)->get()->toArray();
+
+        $expertServiceAcademicFields = DB::select("CALL get_extension_program_fields_by_form_id(3)");
         
-        $collegeAndDepartment = DB::select("CALL get_college_and_department_by_department_id(".$expert_service_in_academic->department_id.")");
+        $documents = ExpertServiceAcademicDocument::where('expert_service_academic_id', $expert_service_in_academic->id)->get()->toArray();
 
-        $classification = DB::select("CALL get_dropdown_name_by_id(".$expert_service_in_academic->classification.")");
-        
-        $nature = DB::select("CALL get_dropdown_name_by_id(".$expert_service_in_academic->nature.")");
+        $values = $expert_service_in_academic->toArray();
+         
 
-        $indexing = DB::select("CALL get_dropdown_name_by_id(".$expert_service_in_academic->indexing.")");
-
-        $level = DB::select("CALL get_dropdown_name_by_id(".$expert_service_in_academic->level.")");
-
-        return view('extension-programs.expert-services.academic.show', compact('expert_service_in_academic', 'expertServiceAcademicDocuments', 
-            'collegeAndDepartment', 'classification', 'nature',
-            'indexing', 'level'));
+        return view('extension-programs.expert-services.academic.show', compact('expertServiceAcademicFields', 'expert_service_in_academic', 'documents', 'values'));
     }
 
     /**
@@ -138,6 +139,8 @@ class AcademicController extends Controller
      */
     public function edit(ExpertServiceAcademic $expert_service_in_academic)
     {
+        $this->authorize('update', ExpertServiceAcademic::class);
+
         if(ExtensionProgramForm::where('id', 3)->pluck('is_active')->first() == 0)
             return view('inactive');
         $expertServiceAcademicFields = DB::select("CALL get_extension_program_fields_by_form_id(3)");
@@ -166,6 +169,7 @@ class AcademicController extends Controller
      */
     public function update(Request $request, ExpertServiceAcademic $expert_service_in_academic)
     {
+        $this->authorize('update', ExpertServiceAcademic::class);
 
         if(ExtensionProgramForm::where('id', 3)->pluck('is_active')->first() == 0)
             return view('inactive');
@@ -176,12 +180,12 @@ class AcademicController extends Controller
             'from' => 'required|date',
             'to' => 'required|date|after_or_equal:from',
             // 'publication_or_audio_visual' => '',
-            // 'copyright_no' => '',
+            'copyright_no' => 'required',
             // 'indexing' => '',
             'level' => 'required',
             'college_id' => 'required',
             'department_id' => 'required',
-            'description' => 'required',
+            // 'description' => 'required',
         ]);
 
         $input = $request->except(['_token', '_method', 'document', 'college_id']);
@@ -222,6 +226,8 @@ class AcademicController extends Controller
      */
     public function destroy(ExpertServiceAcademic $expert_service_in_academic)
     {
+        $this->authorize('delete', ExpertServiceAcademic::class);
+
         if(ExtensionProgramForm::where('id', 3)->pluck('is_active')->first() == 0)
             return view('inactive');
         $expert_service_in_academic->delete();
@@ -230,6 +236,8 @@ class AcademicController extends Controller
     }
 
     public function removeDoc($filename){
+        $this->authorize('delete', ExpertServiceAcademic::class);
+
         if(ExtensionProgramForm::where('id', 3)->pluck('is_active')->first() == 0)
             return view('inactive');
         ExpertServiceAcademicDocument::where('filename', $filename)->delete();
